Last Minute addition, Optional ColorTables for ColorValues

+++ src/main/java/carbonconfiglib/gui/config/ColorElement.java
@@ -72,12 +72,12 @@ public class ColorElement extends ConfigElement
 		super.render(x, top, left, width, height, mouseX, mouseY, selected, partialTicks);
 		if(isArray()) {
 			AbstractGui.fill(left+186, top-1, left+203, top+19, 0xFFA0A0A0);
-			AbstractGui.fill(left+187, top, left+202, top+18, Integer.decode(value.get()) | 0xFF000000);
+			AbstractGui.fill(left+187, top, left+202, top+18, Long.decode(value.get()).intValue() | 0xFF000000);
 		}
 		else {
 			int xOff = isCompound() ? 194 : 207;
 			AbstractGui.fill(left+xOff, top-1, left+xOff+17, top+19, 0xFFA0A0A0);
-			AbstractGui.fill(left+xOff+1, top, left+xOff+16, top+18, Integer.decode(value.get()) | 0xFF000000);
+			AbstractGui.fill(left+xOff+1, top, left+xOff+16, top+18, Long.decode(value.get()).intValue() | 0xFF000000);
 		}
 		if(textBox.isMouseOver(mouseX, mouseY) && result != null && !result.getValue()) {
 			owner.addTooltips(new StringTextComponent(result.getError().getMessage()).applyTextStyle(TextFormatting.RED));
